Extract OnboardingDialog to components, only show if CLOUD
