+++ services/orchest-webserver/client/src/components/OnboardingDialog/assets.jsx
